+++ src/cmake/platform_init.cmake
@@ -1,18 +1,18 @@
 if(CMAKE_SYSTEM_NAME STREQUAL "Windows")
-	set(BNG_IS_WINDOWS TRUE)
-	set(BNG_PLATFORM "win")
+  set(BNG_IS_WINDOWS TRUE)
+  set(BNG_PLATFORM "win")
 elseif(CMAKE_SYSTEM_NAME STREQUAL "Darwin")
-	set(BNG_IS_MACOS TRUE)
-	set(BNG_PLATFORM "mac")
+  set(BNG_IS_MACOS TRUE)
+  set(BNG_PLATFORM "mac")
 elseif(CMAKE_SYSTEM_NAME STREQUAL "Linux")
-	set(BNG_IS_LINUX TRUE)
-	set(BNG_PLATFORM STREQUAL "lnx")
+  set(BNG_IS_LINUX TRUE)
+  set(BNG_PLATFORM STREQUAL "lnx")
 elseif(CMAKE_SYSTEM_NAME STREQUAL "Android")
-	set(BNG_IS_ANDROID TRUE)
-	set(BNG_PLATFORM "and")
+  set(BNG_IS_ANDROID TRUE)
+  set(BNG_PLATFORM "and")
 elseif(CMAKE_SYSTEM_NAME STREQUAL "iOS")
-	set(BNG_IS_IOS TRUE)
-	set(BNG_PLATFORM "ios")
+  set(BNG_IS_IOS TRUE)
+  set(BNG_PLATFORM "ios")
 else()
-	message(FATAL_ERROR "Unsupported platform ${CMAKE_SYSTEM_NAME}")
+  message(FATAL_ERROR "Unsupported platform ${CMAKE_SYSTEM_NAME}")
 endif()
