+++ src/engine/engine.cpp
@@ -1,7 +1,7 @@
 #include "engine.h"
 
 namespace bng::engine {
-	int Engine::foo() const {
-		return 1;
-	}
+  int Engine::foo() const {
+    return 1;
+  }
 }
