+++ src/render/render.cpp
@@ -1,7 +1,7 @@
 #include "render.h"
 
 namespace bng::render {
-	int Render::foo() const {
-		return 3;
-	}
+  int Render::foo() const {
+    return 3;
+  }
 }
