+++ src/render/render.h
@@ -1,10 +1,10 @@
 #pragma once
 
 namespace bng::render {
-	class Render {
-	public:
-		Render() = default;
-		~Render() = default;
-		int foo() const;
-	};
+  class Render {
+  public:
+    Render() = default;
+    ~Render() = default;
+    int foo() const;
+  };
 }
