+++ src/render/tests/test_render.cpp
@@ -1,14 +1,14 @@
 #include "render.h"
-#include "bng_test/bng_test.h"
+#include "test_harness/test_harness.h"
 
 using namespace bng::render;
 
-BNG_TEST(test_1, {
+BNG_TEST(test_a, {
 	Render render;
 	BT_CHECK(render.foo() == 3);
 });
 
-BNG_TEST(test_2, {
+BNG_TEST(test_b, {
 	Render render;
 	BT_CHECK(true);
 	BT_CHECK(render.foo() != 2);
